requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.0
Flask-JWT-Extended>=4.6
Flask-SocketIO>=5.3
scikit-learn>=1.3
flask-bcrypt>=1.0
pypdf>=4.0
python-dotenv>=1.0
SQLAlchemy>=2.0
pytest>=8.0
Flask-Testing>=0.8
gunicorn>=22.0
eventlet>=0.36
psycopg2-binary>=2.9

